LessonSummaryQuiz-main/LessonSummaryAI/requirements.txt
flask
flask-sqlalchemy
flask-login
fpdf
werkzeug
python-docx